refactor: move all middleware to middleware/ directory
